==> 7-biomarker/RMD versions/3_network_analysis.Rmd <==
---
title: |
  | Exploration of Biomarker Data
author: "Sajia Darwish"
date: "`r format(Sys.time(), '%Y-%m-%d')`"
output:
  html_document:
    toc: yes
    toc_depth: 4
  pdf_document:
    toc: yes
    toc_depth: '4'
keep_tex: yes
theme: bclear
fontsize: 10pt
geometry: margin=1in
header-includes:
- \usepackage{enumitem}
- \usepackage{indentfirst}
- \usepackage[default, scale=1]{raleway}
- \usepackage[T1]{fontenc}
- \usepackage{inconsolata}
always_allow_html: yes
---

# ELICIT
```{r, echo=FALSE}
library(tidyverse)
# if (!require("BiocManager", quietly = TRUE))
#     install.packages("BiocManager")
#BiocManager::install("BioNetStat")
library(BioNetStat)
library(igraph)
library(corrr)

hmo <- read.csv("/data/KI/imic/data/raw_lab_data/elicit/milk_analytes/HMO_ELICIT.csv")
hmo <- rename("bmid" = "X", hmo)

sapient <- read.csv("/data/KI/imic/data/raw_lab_data/elicit/milk_analytes/Sapient_ELICIT.csv")
sapient <- rename("bmid_base" = "X", sapient)

bvit <- read.csv("/data/KI/imic/data/raw_lab_data/elicit/milk_analytes/Allen_Bvit_ELICIT.csv")
bvit <- rename("bmid_base" = "X", bvit)

biocratesNorm <- read.csv("/data/KI/imic/data/raw_lab_data/elicit/milk_analytes/Biocrates_ELICIT_Normalized_toALL_NS.csv")
biocratesNorm <- rename("bmid_base" = "X", biocratesNorm)

biocrates <- read.csv("/data/KI/imic/data/raw_lab_data/elicit/milk_analytes/Biocrates_ELICIT_NOT_Normalized.csv")
biocrates <- rename("bmid_base" = "X", biocrates)

metabolInd <- read.csv("/data/KI/imic/data/raw_lab_data/elicit/milk_analytes/MetaboIndicators/Biocrates_ELICIT_Normalized_MetaboINDICATOR.csv")
metabolInd <- rename("bmid_base" = "X", metabolInd)
```

# HMOs: network analysis
```{r, echo=FALSE}
hmo <- hmo %>%
  dplyr::select(ends_with("ug.mL")) %>%
  dplyr::select(!"SUM_ug.mL")

# Cleanup column names
colnames(hmo) <- sub("_.*", "", colnames(hmo))

# Collapse hmo data into long format
long <- pivot_longer(hmo, cols = everything())

# Version 1: simplest network
corMatrix <- cor(hmo, method = "pearson", use = "complete.obs")
corMatrix %>%
  network_plot()

library("Hmisc")
corMatrix2 <- rcorr(as.matrix(hmo))

flattenCorrMatrix <- function(cormat, pmat) {
  ut <- upper.tri(cormat)
  data.frame(
    row = rownames(cormat)[row(cormat)[ut]],
    column = rownames(cormat)[col(cormat)[ut]],
    cor  =(cormat)[ut],
    p = pmat[ut]
    )
}

corMatrix3 <- flattenCorrMatrix(corMatrix2 $ r, corMatrix2 $ P)
corMatrix4 <- corMatrix3 %>%
  select(-p)
```

# Network analysis: first try based on correlation
```{r}
# Edge list
EL <- corMatrix4

EL <- rename(EL, from = row)
EL <- rename(EL, to = column)
EL <- rename(EL, weight = cor)
EL $ weight <- abs(EL $ weight)

# Node list
NL <- as.data.frame(unique(long $ name))

hmoNetwork <- graph_from_data_frame(d = EL, vertices = NL, directed = FALSE)

V(hmoNetwork)
E(hmoNetwork)

plot(hmoNetwork, 
     edge.arrow.size = .5, 
     vertex.color = "plum", 
     vertex.label.dist = 2.5)

E(hmoNetwork) $ width <- E(hmoNetwork) $ weight

plot(hmoNetwork, 
     edge.arrow.size = .7, 
     vertex.label.dist = 2.5, 
     edge.curved = .2)#, 
     #layout = Layout2)

Layout3 <- layout_with_kk(hmoNetwork)

plot(hmoNetwork, 
     edge.arrow.size = 1, 
     vertex.label.dist = 2, 
     layout = Layout3)

## Use correlations between variables "as distance"
dd <- as.dist((1 - cor(hmoScaled)) / 2)
round(1000 * dd) # (prints more nicely)
plot(hclust(dd)) # to see a dendrogram of clustered variables

```

# Network analysis: 2nd try based on distance matrix
```{r}
#test <- scale(hmo, center = F, scale = T) # Not useful for biological data.
# Paper: https://link.springer.com/article/10.1186/1471-2164-7-142

range01 <- function(x, ...){(x - min(x, ...)) / (max(x, ...) - min(x, ...))}
hmoScaled <- range01(hmo)

# Transpose the dataset so that each row is an HMO
transpose <- t(hmoScaled)

distMatrix <- as.matrix(dist(transpose))

g <- graph_from_adjacency_matrix(distMatrix, mode = "undirected")

plot(g, 
     edge.arrow.size = .5,
     edge.color = "gray", 
     vertex.label.color = "darkred", 
     vertex.color = "#cccccc", 
     vertex.label.dist = 1,
     vertex.label.dist = 2)

layout <- layout.circle(g)
plot(g, layout = layout)

layout <- layout.fruchterman.reingold(g)
plot(g, layout = layout)
```
