==> 7-biomarker/RMD versions/4_pca.Rmd <==
---
title: |
  Principal Component Analysis
author: "Sajia Darwish"
date: "`r format(Sys.time(), '%Y-%m-%d')`"
output:
  html_document:
    toc: yes
    toc_depth: 4
  pdf_document:
    toc: yes
    toc_depth: '4'
keep_tex: yes
theme: bclear
fontsize: 10pt
geometry: margin=1in
header-includes:
- \usepackage{enumitem}
- \usepackage{indentfirst}
- \usepackage[default, scale=1]{raleway}
- \usepackage[T1]{fontenc}
- \usepackage{inconsolata}
always_allow_html: yes
---
```{r setup1, include=FALSE}
BV_dir = "/data/KI/imic/"
library(tidyverse)
#library(knitr)
```
Steps taken before implementing PCA:

1) Replaced < LOD with with the minimum observed value for each variable divided by 2.
   Note: after this calculation, some values go to infinity because all observations for some variables were <LOD. They have been set to 0. This doesn't really matter because these variables were later removed due to having zero variance.
2) Checked missingness in the data.
3) Removed zero variance variables.
4) Scaled and centered all variables to ensure that the criterion for finding linear combinations of the predictors is based on how much variation they explain and therefore improve numerical stability.

Notes about the following results:

1) Number of components for each dataset has been chosen based on their respective scree plots.
2) Importance has been calculated based on the contribution of a variable to each component. For example, if all variables would contribute equally to each component they would each take up 1/ncol(data). So, any variable that contributes more than 1/ncol(data) to a component can be considered as an important contributor to that component. The vertical line in each plot represents this threshold.
3) Output data from PCA merged with bmid has been saved to "/data/KI/imic/results/pcaData/".

# ELICIT

## Biocrates normalized: targeted metabolomics data

### Scree plots
```{r, echo=FALSE, fig.align='center', out.width="70%", out.height="70%"}
knitr::include_graphics(paste0(BV_dir,"results/pcaData/elicit/biocratesNormEsPlot.png"))
knitr::include_graphics(paste0(BV_dir,"results/pcaData/elicit/biocratesNormEsPlot2.png"))
```

### Top 10 contributors to each component based on importance
```{r, echo=FALSE, fig.align='center', out.width="70%", out.height="100%"}
knitr::include_graphics(paste0(BV_dir,"results/pcaData/elicit/biocratesNormEtop10cI.png"))
```

### Top 10 contributors to each component by value to see direction
```{r, echo=FALSE, fig.align='center', out.width="70%", out.height="100%"}
knitr::include_graphics(paste0(BV_dir,"results/pcaData/elicit/biocratesNormEtop10cV.png"))
```

## Metabolite indicators: sums and ratios

### Scree plots
```{r, echo=FALSE, fig.align='center', out.width="70%", out.height="70%"}
knitr::include_graphics(paste0(BV_dir,"results/pcaData/elicit/meIndEsPlot.png"))
knitr::include_graphics(paste0(BV_dir,"results/pcaData/elicit/meIndEsPlot2.png"))
```

### Top 10 contributors: importance
```{r, echo=FALSE, fig.align='center', out.width="70%", out.height="100%"}
knitr::include_graphics(paste0(BV_dir,"results/pcaData/elicit/metabolIndtop10cI.png"))
```

### Top 10 contributors: value
```{r, echo=FALSE, fig.align='center', out.width="70%", out.height="100%"}
knitr::include_graphics(paste0(BV_dir,"results/pcaData/elicit/metabolIndtop10cV.png"))
```

# VITAL

## Biocrates normalized: targeted metabolomics data

### Scree plots
```{r, echo=FALSE, fig.align='center', out.width="70%", out.height="70%"}
knitr::include_graphics(paste0(BV_dir,"results/pcaData/vital/biocratesNormVsPlot.png"))
knitr::include_graphics(paste0(BV_dir,"results/pcaData/vital/biocratesNormVsPlot2.png"))
```

### Top 10 contributors: importance
```{r, echo=FALSE, fig.align='center', out.width="70%", out.height="100%"}
knitr::include_graphics(paste0(BV_dir,"results/pcaData/vital/biocratesNormVtop10cI.png"))
```

### Top 10 contributors: value
```{r, echo=FALSE, fig.align='center', out.width="70%", out.height="100%"}
knitr::include_graphics(paste0(BV_dir,"results/pcaData/vital/biocratesNormVtop10cV.png"))
```

## Metabolite indicators: sums and ratios

### Scree plots
```{r, echo=FALSE, fig.align='center', out.width="70%", out.height="70%"}
knitr::include_graphics(paste0(BV_dir,"results/pcaData/vital/metabolIndsPlot.png"))
knitr::include_graphics(paste0(BV_dir,"results/pcaData/vital/metabolIndsPlot2.png"))
```

### Top 10 contributors: importance
```{r, echo=FALSE, fig.align='center', out.width="70%", out.height="100%"}
knitr::include_graphics(paste0(BV_dir,"results/pcaData/vital/metabolIndtop10cI.png"))
```

### Top 10 contributors: value
```{r, echo=FALSE, fig.align='center', out.width="70%", out.height="100%"}
knitr::include_graphics(paste0(BV_dir,"results/pcaData/vital/metabolIndtop10cV.png"))
```

## Proteomics data

### Scree plots
```{r, echo=FALSE, fig.align='center', out.width="70%", out.height="70%"}
knitr::include_graphics(paste0(BV_dir,"results/pcaData/vital/pblNormVsPlot.png"))
knitr::include_graphics(paste0(BV_dir,"results/pcaData/vital/pblNormVsPlot2.png"))
```

### Top 10 contributors: importance
```{r, echo=FALSE, fig.align='center', out.width="70%", out.height="100%"}
knitr::include_graphics(paste0(BV_dir,"results/pcaData/vital/pblNormVtop10cI.png"))
```

### Top 10 contributors: value
```{r, echo=FALSE, fig.align='center', out.width="70%", out.height="100%"}
knitr::include_graphics(paste0(BV_dir,"results/pcaData/vital/pblNormVtop10cV.png"))
```
